+++ packages/polaris-viz/src/components/BarChart/index.ts
@@ -1,9 +1,4 @@
 export {BarChart} from './BarChart';
 export type {BarChartProps} from './BarChart';
-export type {
-  AnnotationLookupTable,
-  RenderTooltipContentData,
-  XAxisOptions,
-  YAxisOptions,
-} from './types';
+export type {AnnotationLookupTable, RenderTooltipContentData} from './types';
 export {AnnotationLine} from './components/AnnotationLine';