+++ packages/polaris-viz/src/components/VerticalBarChart/VerticalBarChart.tsx
@@ -1,13 +1,16 @@
 import React from 'react';
-import type {ChartType, DataSeries} from '@shopify/polaris-viz-core';
+import type {
+  ChartType,
+  DataSeries,
+  XAxisOptions,
+  YAxisOptions,
+} from '@shopify/polaris-viz-core';
 
 import {ChartContainer} from '../../components/ChartContainer';
 import {useTheme, useThemeSeriesColors} from '../../hooks';
 import type {
   AnnotationLookupTable,
   RenderTooltipContentData,
-  XAxisOptions,
-  YAxisOptions,
 } from '../../components/BarChart';
 
 import {Chart} from './Chart';