+++ packages/polaris-viz/src/components/VerticalBarChart/hooks/use-y-scale.ts
@@ -1,10 +1,10 @@
 import {useMemo} from 'react';
 import {scaleLinear} from 'd3-scale';
-import type {DataSeries} from '@shopify/polaris-viz-core';
+import type {DataSeries, LabelFormatter} from '@shopify/polaris-viz-core';
 
 import {shouldRoundScaleUp, getStackedMinMax} from '../../../utilities';
 import {MIN_Y_LABEL_SPACE} from '../constants';
-import type {NumberLabelFormatter, StackedSeries} from '../../../types';
+import type {StackedSeries} from '../../../types';
 
 export function useYScale({
   drawableHeight,
@@ -15,7 +15,7 @@ export function useYScale({
 }: {
   drawableHeight: number;
   data: DataSeries[];
-  formatYAxisLabel: NumberLabelFormatter;
+  formatYAxisLabel: LabelFormatter;
   stackedValues: StackedSeries[] | null;
   integersOnly: boolean;
 }) {