+++ packages/polaris-viz/CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- All components now use the core `XAxisOptions` & `YAxisOptions` types.
+- All label formatters now use the core `LabelFormatter` type.
+
+### Removed
+
+- Removed `xAxisLabels` from `LinearXAxisOptions.xAxisOptions`.
+- Removed `StringLabelFormatter` & `NumberLabelFormatter` types.
 
 ## [1.1.0] - 2022-03-28